fix(server): finalize modular refactor
